+++ mmcls/models/utils/__init__.py
@@ -1,10 +1,10 @@
+from .augment.cutmix import BatchCutMixLayer
+from .augment.mixup import BatchMixupLayer
 from .channel_shuffle import channel_shuffle
-from .cutmix import BatchCutMixLayer
 from .cutmixup import CutMixUp
 from .helpers import to_2tuple, to_3tuple, to_4tuple, to_ntuple
 from .inverted_residual import InvertedResidual
 from .make_divisible import make_divisible
-from .mixup import BatchMixupLayer
 from .se_layer import SELayer
 
 __all__ = [
